project/mlops_project/web_service_mlflow/requirements.txt
flask
gunicorn
pandas
boto3
python-dotenv
mlflow 
autogluon.tabular[all]